requirements.txt
asgiref==3.6.0
Django==3.2.18
django-bootstrap-datepicker-plus==5.0.3
django_debug_toolbar==3.8.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dotenv==0.21.1
pytz==2023.3
sorl-thumbnail==12.9.0
sqlparse==0.4.3
typing_extensions==4.5.0
